+++ src/pages/Home.jsx
@@ -1,102 +1,19 @@
-import React, { useState } from "react";
-import ImageList from "@mui/material/ImageList";
-import ImageListItem from "@mui/material/ImageListItem";
-import ImageListItemBar from "@mui/material/ImageListItemBar";
-import IconButton from "@mui/material/IconButton";
-import StarBorderIcon from "@mui/icons-material/StarBorder";
-// import { useDispatch, useSelector } from "react-redux";
-import SearchBar from "../components/SearchBar/SearchBar";
-// import CircularProgress from "@mui/material/CircularProgress";
+import React from "react";
+import ImageLists from "../components/ImageList/ImageList";
 
-function srcset(image, width, height, rows = 1, cols = 1) {
-  return {
-    src: `${image}?w=${width * cols}&h=${height * rows}&fit=crop&auto=format`,
-    srcSet: `${image}?w=${width * cols}&h=${
-      height * rows
-    }&fit=crop&auto=format&dpr=2 2x`,
-  };
-}
+
+import SearchBar from "../components/SearchBar/SearchBar";
+
+
+
 const Home = () => {
-  // const dispatch = useDispatch();
-  // const {images} = useSelector((state)=> state.selectImage)
-  // const [search, setSearch] = useState(false);
+
   return (
     <>
       <SearchBar />
-      <ImageList
-        sx={{
-          width: 300,
-          height: 550,
-          padding: 1,
-          transform: "translateZ(0)",
-        }}
-        rowHeight={200}
-        gap={1}
-      >
-        {itemData.map((item) => {
-          const cols = item.featured ? 2 : 1;
-          const rows = item.featured ? 2 : 1;
-
-          return (
-            <ImageListItem key={item.img} cols={cols} rows={rows}>
-              <img
-                {...srcset(item.img, 250, 200, rows, cols)}
-                alt="title"
-                loading="lazy"
-              />
-              <ImageListItemBar
-                sx={{
-                  background:
-                    "linear-gradient(to bottom, rgba(0,0,0,0.7) 0%, " +
-                    "rgba(0,0,0,0.3) 70%, rgba(0,0,0,0) 100%)",
-                }}
-                position="top"
-                actionIcon={
-                  <IconButton sx={{ color: "white" }}>
-                    <StarBorderIcon />
-                  </IconButton>
-                }
-                actionPosition="left"
-              />
-            </ImageListItem>
-          );
-        })}
-      </ImageList>
-    </>
+      <ImageLists/>
+      </>
   );
 };
-const itemData = [
-  {
-    img: "https://images.unsplash.com/photo-1551963831-b3b1ca40c98e",
-    title: "Breakfast",
-    author: "@bkristastucchio",
-    featured: true,
-  },
-  {
-    img: "https://images.unsplash.com/photo-1551782450-a2132b4ba21d",
-    title: "Burger",
-    author: "@rollelflex_graphy726",
-  },
-  {
-    img: "https://images.unsplash.com/photo-1522770179533-24471fcdba45",
-    title: "Camera",
-    author: "@helloimnik",
-  },
-  {
-    img: "https://images.unsplash.com/photo-1444418776041-9c7e33cc5a9c",
-    title: "Coffee",
-    author: "@nolanissac",
-  },
-  {
-    img: "https://images.unsplash.com/photo-1533827432537-70133748f5c8",
-    title: "Hats",
-    author: "@hjrc33",
-  },
-  {
-    img: "https://images.unsplash.com/photo-1558642452-9d2a7deb7f62",
-    title: "Honey",
-    author: "@arwinneil",
-    featured: true,
-  },
-];
+
 export default Home;
